Move migrateTUI to migrate, replacing old tool
