+++ ingesters/migrate/.gitignore
@@ -1,1 +1,2 @@
-migrate
+migrateTUI
+test.conf
