+++ ingesters/migrate/main.go
@@ -2,7 +2,7 @@
 // +build linux
 
 /*************************************************************************
- * Copyright 2017 Gravwell, Inc. All rights reserved.
+ * Copyright 2022 Gravwell, Inc. All rights reserved.
  * Contact: <legal@gravwell.io>
  *
  * This software may be modified and distributed under the terms of the
@@ -14,9 +14,9 @@ package main
 import (
 	"context"
 	"flag"
-	"fmt"
 	"net"
 	"os"
+	"time"
 
 	"github.com/gravwell/gravwell/v3/ingest"
 	"github.com/gravwell/gravwell/v3/ingest/config/validate"
@@ -34,9 +34,16 @@ var (
 	v        bool
 	lg       *log.Logger
 	src      net.IP
+
+	st *StateTracker
+
+	igst *ingest.IngestMuxer
+
+	cfg *cfgType
 )
 
 func init() {
+	v = true
 	flag.Parse()
 	if *ver {
 		version.PrintVersion(os.Stdout)
@@ -44,92 +51,73 @@ func init() {
 		os.Exit(0)
 	}
 	lg = log.New(&discard{})
-	if !*status {
-		lg.AddWriter(os.Stdout)
-	}
+	//lg.AddWriter(os.Stderr)
 	lg.SetAppname(appName)
 	validate.ValidateConfig(GetConfig, *confLoc, *confdLoc)
 }
 
 func main() {
+	// Make a local writer so we can write to the console if something goes wrong
+	llg := log.New(&discard{})
+	llg.AddWriter(os.Stderr)
+	llg.SetAppname(appName)
+
+	var err error
+	doneChan := make(chan bool)
+	time.Sleep(500 * time.Millisecond)
 	// this thing hits the filesystem, parallelism will almost always be bad
 	utils.MaxProcTune(1)
-	cfg, err := GetConfig(*confLoc, *confdLoc)
+	cfg, err = GetConfig(*confLoc, *confdLoc)
 	if err != nil {
-		lg.FatalCode(0, "failed to get configuration", log.KVErr(err))
+		llg.FatalCode(0, "failed to get configuration", log.KVErr(err))
 	}
 
 	if len(cfg.Log_File) > 0 {
 		fout, err := os.OpenFile(cfg.Log_File, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0640)
 		if err != nil {
-			lg.FatalCode(0, "failed to open log file", log.KV("path", cfg.Log_File), log.KVErr(err))
+			llg.FatalCode(0, "failed to open log file", log.KV("path", cfg.Log_File), log.KVErr(err))
 		}
 		if err = lg.AddWriter(fout); err != nil {
-			lg.Fatal("failed to add a writer", log.KVErr(err))
+			llg.Fatal("failed to add a writer", log.KVErr(err))
+		}
+		if err = llg.AddWriter(fout); err != nil {
+			llg.Fatal("failed to add a writer", log.KVErr(err))
 		}
 		if len(cfg.Log_Level) > 0 {
 			if err = lg.SetLevelString(cfg.Log_Level); err != nil {
-				lg.FatalCode(0, "invalid Log Level", log.KV("loglevel", cfg.Log_Level), log.KVErr(err))
+				llg.FatalCode(0, "invalid Log Level", log.KV("loglevel", cfg.Log_Level), log.KVErr(err))
 			}
 		}
 	}
 
-	igst := getIngestConnection(cfg, lg)
-	st, err := NewStateTracker(cfg.StatePath())
+	ctx, _ := context.WithCancel(context.Background())
+	st, err = NewStateTracker(cfg.StatePath())
 	if err != nil {
-		lg.FatalCode(0, "Failed to load state store file", log.KVErr(err))
-	}
-	ctx, cf := context.WithCancel(context.Background())
-	qc := utils.GetQuitChannel()
-	statusChan, errChan := processSources(cfg, st, igst, ctx)
-	if *status {
-		go statusPrinter(statusChan)
-	} else {
-		go statusEater(statusChan)
+		llg.FatalCode(0, "Failed to load state store file", log.KVErr(err))
 	}
 
-	select {
-	case <-qc:
-		lg.Info("interrupt caught, waiting for processors to exit")
-		fmt.Println("interrupt caught, waiting for processors to exit")
-		cf()
-		err = <-errChan //wait for the processor to exit
-		debugout("Finished with %v after interrupt\n", err)
-	case err = <-errChan:
-		cf()
-		if err != nil {
-			lg.Error("migration error", log.KVErr(err))
-			debugout("error: %v\n", err)
-		} else {
-			debugout("done\n")
-		}
-	}
+	go guiMain(doneChan, st)
 
-	debugout("Closing migration engine\n")
-	if err = igst.Close(); err != nil {
-		st.Close()
-		lg.FatalCode(0, "failed to close ingest connection", log.KVErr(err))
-	} else if err = st.Close(); err != nil {
-		lg.FatalCode(0, "failed to close state store", log.KVErr(err))
-	}
-	debugout("Completed\n")
-}
-
-func processSources(cfg *cfgType, st *StateTracker, igst *ingest.IngestMuxer, ctx context.Context) (<-chan statusUpdate, <-chan error) {
-	sc := make(chan statusUpdate, 1)
-	ec := make(chan error, 1)
-	go processSourcesAsync(cfg, st, igst, ctx, sc, ec)
-	return sc, ec
-}
-
-func processSourcesAsync(cfg *cfgType, st *StateTracker, igst *ingest.IngestMuxer, ctx context.Context, sc chan statusUpdate, errch chan error) {
-	defer close(errch)
-	defer close(sc)
-	//process flat files
-	if err := processFiles(cfg, st, igst, ctx, sc); err != nil {
-		errch <- err
+	// Set up the early Splunk stuff
+	if stop, err := initializeSplunk(cfg, st, ctx); err != nil {
+		llg.FatalCode(0, "Failed to initialize splunk", log.KVErr(err))
+	} else if stop {
 		return
 	}
 
-	return
+	igst = getIngestConnection(cfg, lg)
+
+	<-doneChan
+
+	// write out the statuses to the state file
+	for _, v := range splunkTracker.GetAllStatuses() {
+		st.Add(splunkStateType, v)
+	}
+
+	if err = igst.Close(); err != nil {
+		st.Close()
+		llg.FatalCode(0, "failed to close ingest connection", log.KVErr(err))
+	} else if err = st.Close(); err != nil {
+		llg.FatalCode(0, "failed to close state store", log.KVErr(err))
+	}
 }
