--- ingesters/migrate/migrate.conf
@@ -1,23 +1,0 @@
-[Global]
-Ingest-Secret = "IngestSecrets"
-Connection-Timeout = 0
-Insecure-Skip-TLS-Verify=false
-#Cleartext-Backend-Target=127.0.0.1:4023 #example of adding a cleartext connection
-#Encrypted-Backend-Target=127.1.1.1:4024 #example of adding an encrypted connection
-#Pipe-Backend-Target=/opt/gravwell/comms/pipe #a named pipe connection, this should be used when ingester is on the same machine as a backend
-State-Store-Location=/tmp/migrate.state
-Log-Level=INFO #options are OFF INFO WARN ERROR
-Log-File=/tmp/migrate.log
-
-#basic default logger, all entries will go to the default tag
-#no Tag-Name means use the default tag
-[Files "auth"]
-	Base-Directory="/var/log/"
-	File-Filter="auth.log,auth.log.[0-9]" #we are looking for all authorization log files
-	Tag-Name=auth
-	Assume-Local-Timezone=true #Default for assume localtime is false
-#	Ignore-Timestamps=true #do not ignore timestamps
-#	Recursive=true
-#	Ignore-Line-Prefix="#" # ignore lines beginning with #
-#	Ignore-Line-Prefix="//"
-#	Timezone-Override="UTC" #force the timezone
