--- ingesters/migrateTUI/.gitignore
@@ -1,2 +1,0 @@
-migrateTUI
-test.conf
